pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "geospeed"
version = "0.1.0"
description = "Speed comparison of modern GIS tools for spatial overlay operations"
authors = [{name = "SeHe"}]
requires-python = ">=3.12.11"
dependencies = [
    "geopandas==1.0.1",
    "pyogrio==0.9.0",
    "fiona==1.10.1",
    "pyarrow",
    "dask-geopandas",
    "matplotlib",
    "memory-profiler",
    "psutil",
    "duckdb~=1.1.0",
    "pyspark==3.4.1",
    "apache-sedona[spark]==1.6.1",
    # "geofileops>=0.8.0",  # Requires system GDAL, installed in CI
]

[dependency-groups]
dev = [
    "mypy",
    "ruff",
    "ty>=0.0.1a19",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".eggs",
    ".git",
    ".venv",
    ".ipynb_checkpoints"
]
include = ["*.py", "*.ipynb"]

line-length = 120
indent-width = 4
target-version = "py312"

lint.select = ["ALL"]
lint.ignore = ["D203", "D212", "ISC001", "COM812", "E501",
                "T201"]

# Allow fix for all enabled rules (when `--fix`) is provided.
lint.fixable = ["ALL"]
lint.unfixable = []

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
